+++ Proyecto1_SO/src/Workers/Director.java
@@ -76,7 +76,7 @@ public class Director extends Worker {
             }
             
             this.getDrive().getCostsM().acquire();
-            this.getDrive().setCostDirector(this.getDrive().getCostDirector()+this.getHourlyRate()*24);
+            this.getDrive().setCostDirector((int)this.getDrive().getCostDirector()+this.getHourlyRate()*24);
             this.getDrive().getCostsM().release();
 
         } catch (Exception e) {
